--- tests/test_suite.cpp
@@ -1,3 +1,0 @@
-#define CATCH_CONFIG_MAIN // This tells Catch to provide a main() - only do this in one cpp file
-#define CATCH_CONFIG_CONSOLE_WIDTH 120
-#include "Catch2/catch.hpp"
